fix accrual

+++ internal/app/gophermart/dto/balance.go
@@ -2,5 +2,5 @@ package dto
 
 type BalanceResponce struct {
 	Current   float32 `json:"current" validate:"required"`
-	Withdrawn int     `json:"withdrawn" validate:"required"`
+	Withdrawn float32 `json:"withdrawn" validate:"required"`
 }
